requirements/base.txt
# Django & DRF
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.2

# Database
mysqlclient==2.2.1
psycopg2-binary==2.9.9  # Alternative: PostgreSQL

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Security
django-cors-headers==4.3.1
django-ratelimit==4.1.0
cryptography==42.0.0
PyJWT==2.8.0

# Password hashing
bcrypt==4.1.2
argon2-cffi==23.1.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# File uploads & storage
Pillow
boto3==1.34.34  # AWS S3
django-storages==1.14.2

# Phone number validation
phonenumbers==8.13.27
django-phonenumber-field==7.3.0

# Face recognition (KYC)
# face-recognition==1.3.0
# dlib==19.24.2
# opencv-python==4.9.0.80

# HTTP requests
requests==2.31.0
httpx==0.26.0

# Data validation
pydantic==2.5.3
marshmallow==3.20.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
shortuuid==1.0.11

# Monitoring & Logging
sentry-sdk==1.40.0
django-extensions==3.2.3

# API documentation
# drf-yasg==1.21.7  # Swagger/OpenAPI
# setuptools # for python > 3.12
drf-spectacular